refactor: split theming module into definitions & engine components
