+++ src/ui/display/ascii_art.py
@@ -6,7 +6,7 @@ from __future__ import annotations
 from pathlib import Path
 from ..core.rich_components import Text
 from ...loom_io.console import console
-from ..theming.colors import get_active_theme, natural_gradient
+from ..theming.theme_engine import get_active_theme, natural_gradient
 
 # * Display the Loom banner w/ smooth gradient styling  
 def show_loom_art(theme_colors: list[str] | None = None) -> None:
